+++ src/org/ods/orchestration/util/DocumentHistory.groovy
@@ -255,9 +255,9 @@ class DocumentHistory {
     }
 
     @NonCPS
-    private static List<String> getConcurrentVersions(List<Map> versions, String previousProjVersion) {
-        // DO NOT remove this method. Takewhile is not supported by Jenkins and must be used in a Non-CPS method
-        versions.takeWhile { it.version != previousProjVersion }.collect { it.id }
+    private static List<Map> getConcurrentVersions(List<Map> versions, String previousProjVersion) {
+        return versions
+            .findAll { it.previousVersion == previousProjVersion }
     }
 
     @NonCPS
@@ -334,8 +334,7 @@ class DocumentHistory {
         } else {
             def pluralS = (concurrentVersions.size() == 1) ? '' : 's'
             return " This document version invalidates the previous document version${pluralS} " +
-                "'${currentEntry.getProjectVersion()}/" +
-                "${concurrentVersions.join("', '${currentEntry.getProjectVersion()}/")}'."
+                concurrentVersions.toSorted{it.id}.reverse().collect{ v -> "'${v.version}/${v.id}'" }.join(', ') + "."
         }
     }