+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Fix branch mapping logic to ensure correct environment assignment ([#1263](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1263))
 * Fix Nexus not accepting npm artifacts when uploaded through odsComponentStageUploadToNexus ([#1268](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1268))
 * Aqua Stage is being skipped when branches are not eligable ([#1170](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1170))
+* Fix partial TIR document history rational to invalidate only existing versions when previous project version is missing
 
 ## [4.12.1] - 2026-04-10
 ### Fixed