+++ CHANGELOG.md
@@ -11,7 +11,10 @@
 * Fix branch mapping logic to ensure correct environment assignment ([#1263](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1263))
 * Fix Nexus not accepting npm artifacts when uploaded through odsComponentStageUploadToNexus ([#1268](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1268))
 * Aqua Stage is being skipped when branches are not eligable ([#1170](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1170))
-* Fix referenced document version incorrectly incremented in getVersion() by preferring Bitbucket history over Jira ([#1274] https://github.com/opendevstack/ods-jenkins-shared-library/pull/1272)
+
+## [4.12.2] - 2026-06-04
+### Fixed
+* Fix referenced document version incorrectly incremented in getVersion() by preferring Bitbucket history over Jira ([#1272] https://github.com/opendevstack/ods-jenkins-shared-library/pull/1272)
 * Fix document history rational to invalidate only concurrent versions ([#1273](https://github.com/opendevstack/ods-jenkins-shared-library/pull/1273))
 
 ## [4.12.1] - 2026-04-10